VDB Parsing Demo
Discovered with others in the 3DS Edition Discord that NBT is actually stored in these, and the keys are there too. So these sound like they are their own implementation, and they aren't implemented elsewhere, it's up to use to figure out the parsing for it.

```
CDB Contains all NBT Data of Blocks, and block places, pos, facing, y, z, x.
VDB Contains all Player Info and other stuff and is a Binary format itself.
Both are Derivitives of LDB, that are both proprietary, and do not act like any version of released LDB to date in the slightest.
```

I noticed that they don't seem to use any compression, so I tried parsing the bytes as NBT after having looked at the values in a hex editor, and parsing from this offset works! Now I can probably write a parser for the header, and make it dynamic to the key string at the start.

```
VDB Files that are 0x10014 Bytes hold GameRules and current stuff such as BiomeData or some structures.
VDB Files that are 0x8014 Bytes hold logs and some mob info
VDB FIles that are 0x200014 Bytes hold Player Info, such as inv data, pos, health, map exploration, etc
```