VDB Initial Setup
More of a restructure, this needs a proper universal implementation for all VDB files, not just this demo one I am testing on. The values and offsets in them will change depending on what the underlying data looks like.

This was from earlier 'yesterday', I just didn't get to check/commit it until now. Wanted to make sure what I committed was how I wanted it to be set up.

Was looking into the `return await` dilemma, still not sure which way to go now, seems like the new modern conclusion is that it's okay, and maybe even better to use the `await` version. It may help with stack traces too.